working/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
pyotp==2.9.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
redis==5.0.1
slowapi==0.1.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
